+++ fj-doc-base-yaml/src/main/java/org/fugerit/java/doc/yaml/parse/DocJsonToYaml.java
@@ -1,0 +1,11 @@
+package org.fugerit.java.doc.yaml.parse;
+
+import org.fugerit.java.doc.json.parse.JsonConstants;
+
+public class DocJsonToYaml extends DocYamlToJson {
+
+    public DocJsonToYaml() {
+        super(JsonConstants.getDefaultMapper(), YamlConstants.getDefaultMapper());
+    }
+
+}