+++ fj-doc-base-yaml/src/main/java/org/fugerit/java/doc/yaml/parse/DocXmlToYaml.java
@@ -1,0 +1,11 @@
+package org.fugerit.java.doc.yaml.parse;
+
+import org.fugerit.java.doc.json.parse.DocXmlToJson;
+
+public class DocXmlToYaml extends DocXmlToJson {
+
+    public DocXmlToYaml() {
+        super( YamlConstants.getDefaultMapper() );
+    }
+
+}