+++ fj-doc-base-yaml/src/main/java/org/fugerit/java/doc/yaml/parse/DocYamlToXml.java
@@ -3,12 +3,11 @@ package org.fugerit.java.doc.yaml.parse;
 import org.fugerit.java.doc.json.parse.DocJsonToXml;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 
 public class DocYamlToXml extends DocJsonToXml {
 
 	public DocYamlToXml() {
-		super( new ObjectMapper( new YAMLFactory() ) );
+		super( YamlConstants.getDefaultMapper() );
 	}
 
 	public DocYamlToXml(ObjectMapper mapper) {