+++ fj-doc-base/src/test/java/test/org/fugerit/java/doc/base/facade/TestDocFacadeSource.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.StringReader;
+import java.io.StringWriter;
 
 import org.fugerit.java.core.cfg.ConfigRuntimeException;
 import org.fugerit.java.core.function.SafeFunction;
@@ -88,6 +89,42 @@ class TestDocFacadeSource {
 		assertEquals( input, DocFacadeSource.cleanSource( input, DocFacadeSource.SOURCE_TYPE_XML ) );
 		assertEquals( input, DocFacadeSource.cleanSource( input, DocFacadeSource.SOURCE_TYPE_JSON ) );
 		assertEquals( input, DocFacadeSource.cleanSource( input, DocFacadeSource.SOURCE_TYPE_YAML ) );
+	}
+
+	private static final DocFacadeSource DOC_FACADE_SOURCE = DocFacadeSource.getInstance();
+
+	@Test
+	void testFromSourceNotSupported() throws IOException {
+		try ( StringReader from = new StringReader( "{}" );
+			StringWriter to = new StringWriter() ) {
+			assertThrows( ConfigRuntimeException.class,
+					() -> DOC_FACADE_SOURCE.convert( from, DocFacadeSource.SOURCE_TYPE_JSON, to, DocFacadeSource.SOURCE_TYPE_XML ) );
+		}
+	}
+
+	@Test
+	void testToSourceNotSupported() throws IOException {
+		try ( StringReader from = new StringReader( "<doc/>" );
+			  StringWriter to = new StringWriter() ) {
+			assertThrows( ConfigRuntimeException.class,
+					() -> DOC_FACADE_SOURCE.convert( from, DocFacadeSource.SOURCE_TYPE_XML, to, DocFacadeSource.SOURCE_TYPE_YAML ) );
+		}
+	}
+
+	@Test
+	void testDocConvertNotFound() {
+		assertThrows( ConfigRuntimeException.class,
+				() -> DocFacadeSource.findDocConvert( DocFacadeSource.SOURCE_TYPE_XML, Integer.MAX_VALUE ) );
+	}
+
+	@Test
+	void testReflectiveConversionXml() throws IOException {
+		String input = "<doc></doc>";
+		try ( StringReader from = new StringReader( input );
+			  StringWriter to = new StringWriter() ) {
+			DocFacadeSource.getInstance().convert( from, DocFacadeSource.SOURCE_TYPE_XML, to, DocFacadeSource.SOURCE_TYPE_XML );
+			Assertions.assertEquals( input, to.toString() );
+		}
 	}
 
 }