+++ fj-doc-maven-plugin/src/test/java/test/org/fugerit/java/doc/project/facade/TestAddVenusFacade.java
@@ -205,6 +205,30 @@ class TestAddVenusFacade {
     }
 
     @Test
+    void testMojoCoreHandlers() throws IOException, MojoExecutionException, MojoFailureException {
+        for ( String currentConfig : Arrays.asList( "ok3-pom" ) ) {
+            File projectDir = this.initConfigWorker( currentConfig );
+            MojoAdd mojoAdd = new MojoAdd() {
+                @Override
+                public void execute() throws MojoExecutionException, MojoFailureException {
+                    this.version = "8.16.6";
+                    this.extensions = "fj-doc-base,fj-doc-base-json,fj-doc-base-yaml,fj-doc-freemarker,fj-doc-mod-fop,fj-doc-mod-poi,fj-doc-mod-opencsv";
+                    this.projectFolder = projectDir.getAbsolutePath();
+                    this.addDocFacade = true;
+                    this.force = true;
+                    this.excludeXmlApis = true;
+                    this.addVerifyPlugin = true;
+                    this.addJunit5 = true;
+                    this.addLombok = true;
+                    super.execute();
+                }
+            };
+            mojoAdd.execute();
+            Assertions.assertTrue( projectDir.exists() );
+        }
+    }
+
+    @Test
     void testFjVersionCheck() {
         Model model = new Model();
         String projectPomFjCoreVersion = BasicVenusFacade.versionToCheck( FJCoreMaven.FJ_CORE_GROUP_ID, FJCoreMaven.FJ_CORE_ARTIFACT_ID, model );