+++ fj-doc-maven-plugin/src/test/java/test/org/fugerit/java/doc/project/facade/TestVenusContext.java
@@ -18,12 +18,19 @@ class TestVenusContext {
         Assertions.assertNotNull( context.getTestResourcesFolder() );
         Assertions.assertNotNull( context.getMainResourcesFolder() );
         Assertions.assertFalse( context.isAsciidocFreemarkerHandlerAvailable() );
+        Assertions.assertTrue( context.isCoreHandlersNotAvailable() );
     }
 
     @Test
     void testAsciiDocCheck() {
         VenusContext context = new VenusContext( TARGET, "8.10.8", "fj-doc-base" );
         Assertions.assertTrue( context.isAsciidocFreemarkerHandlerAvailable() );
+    }
+
+    @Test
+    void testVenusContextCoreHandlers() {
+        VenusContext context = new VenusContext( TARGET, "8.17.0", "fj-doc-base" );
+        Assertions.assertFalse( context.isCoreHandlersNotAvailable() );
     }
 
 }